Moved the inputs into a input folder
